% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backsolve.R
\name{backsolveBatch}
\alias{backsolveBatch}
\title{backsolveBatch}
\usage{
backsolveBatch(
  C,
  A,
  B,
  numbatchB,
  diagIsOne,
  Nglobal,
  Nlocal,
  NlocalCache,
  Cstartend,
  Astartend,
  Bstartend,
  verbose = FALSE
)
}
\arguments{
\item{C}{an object of class 'vclMatrix'}

\item{A}{an object of class 'vclMatrix', upper triangular values are 0}

\item{B}{an object of class 'vclMatrix', batches of rectangular matrices}

\item{numbatchB}{number of batches in B, if 1 then B has the same matrix for all matrices in A batch}

\item{diagIsOne}{logical, indicates if all the diagonal entries in A matrices are 1}

\item{Nglobal}{Size of the index space for use}

\item{Nlocal}{Work group size of the index space}

\item{NlocalCache}{local memory cached}

\item{Cstartend}{a vector that selects the range of C, c(startrow, numberofrows, startcolumn, numberofcols), row starts from 0}

\item{Astartend}{a vector that selects the range of A}

\item{Bstartend}{a vector that selects the range of B}

\item{verbose}{if TRUE, print out more information}
}
\description{
solve A * C = B for C on a GPU, where A, B, C are batches of square matrices
}
\note{
result matrices are stored in C respectively, no returned objects
}
