% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configParams.R
\name{configParams}
\alias{configParams}
\title{set loglikelihood locations}
\usage{
configParams(
  Model,
  alpha = c(0.001, 0.01, 0.1, 0.2, 0.5, 0.8, 0.95, 0.99),
  alphasecond = NULL,
  Mle = NULL,
  boxcox = NULL,
  shapeRestrict = 1000
)
}
\description{
set loglikelihood locations
}
