% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprodBatch.R
\name{crossprodBatch}
\alias{crossprodBatch}
\title{crossprodBatch}
\usage{
crossprodBatch(
  C,
  A,
  D,
  invertD,
  Nglobal,
  Nlocal,
  NlocalCache,
  Cstartend,
  Astartend,
  Dstartend
)
}
\arguments{
\item{C}{a vclMatrix on GPU, square matrices batches}

\item{A}{a vclMatrix on GPU, rectangular matrix batches}

\item{D}{a vclMatrix on GPU, rectangular matrix batches, rows are diagonals of D', stacked row-wise}

\item{invertD}{set to 1 for C = t(A) D^(-1) A}

\item{Nglobal}{vector of number of global work items}

\item{Nlocal}{vector of number of local work items}

\item{NlocalCache}{elements in local cache}

\item{Cstartend}{a vector (startrow, nRow, startcol,nCol) that indicates the selected part of each submatrix in C}

\item{Astartend}{a vector (startrow, nRow, startcol,nCol) that indicates the selected part of each submatrix in A}

\item{Dstartend}{a vector (startrow, nRow, startcol,nCol) that indicates the selected part of each submatrix in D}
}
\description{
computes C = t(A) A or t(A) D' A or t(A) D'^(-1) A in batches on a GPU
}
\note{
computed results are stored in C, no returned objects
}
