% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossprodBatchBackend}
\alias{crossprodBatchBackend}
\title{Multiply crossproduct matrices}
\usage{
crossprodBatchBackend(
  C,
  A,
  D,
  invertD,
  Cstartend,
  Astartend,
  Dstartend,
  Nglobal,
  Nlocal,
  NlocalCache,
  verbose
)
}
\description{
Computes C = t(A) D A
}
