% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfLogL.R
\name{getProfLogL}
\alias{getProfLogL}
\title{gpu elements}
\usage{
getProfLogL(
  data,
  formula,
  coordinates,
  params,
  boxcox,
  type = c("float", "double")[1 + gpuInfo()$double_support],
  NparamPerIter,
  gpuElementsOnly = FALSE,
  reml = FALSE,
  Nglobal,
  Nlocal,
  NlocalCache,
  verbose = c(1, 0)
)
}
\description{
gpu elements
}
