% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLgmCov1d.R
\name{likfitLgmCov1d}
\alias{likfitLgmCov1d}
\title{Estimate profile Log-likelihood for covariance parameters and lambda}
\usage{
likfitLgmCov1d(
  data,
  formula,
  coordinates,
  params,
  paramToEstimate = c("range", "nugget"),
  boxcox,
  cilevel,
  type = c("float", "double")[1 + gpuInfo()$double_support],
  reml = FALSE,
  NparamPerIter,
  Nglobal,
  Nlocal,
  NlocalCache,
  verbose = FALSE
)
}
\description{
Estimate profile Log-likelihood for covariance parameters and lambda
}
