% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prof1dBetas.R
\name{makeSymm}
\alias{makeSymm}
\title{Estimate Log-likelihood for Gaussian random fields when Betas are given}
\usage{
makeSymm(m)
}
\description{
Estimate Log-likelihood for Gaussian random fields when Betas are given
}
