% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternParamsGpu.R
\name{maternGpuParam}
\alias{maternGpuParam}
\title{maternGpuParam}
\usage{
maternGpuParam(x, type = "double")
}
\arguments{
\item{x}{R matrix of covariance parameters}

\item{type}{precision type of paramters, 'double' or 'float'}
}
\description{
Create parameters be used by maternBatch on GPU
}
