% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiplyLowerBatchBackend}
\alias{multiplyLowerBatchBackend}
\title{Multiply lower triangular matrices}
\usage{
multiplyLowerBatchBackend(C, A, B, diagIsOne, Nglobal, Nlocal, NlocalCache)
}
\arguments{
\item{C}{output matrices, stacked row-wise}

\item{A}{lower triangular matrices}

\item{B}{rectangular matrix or matrices}

\item{Nglobal}{vector of number of global work items: Drow, Dcol, Dmatrix}

\item{Nlocal}{vector of number of local work items anything, anything, 1}

\item{NlocalCache}{elements in local cache}
}
\description{
Multiplies a lower triangular matrix by a rectangular matrix
}
