% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyLowerDiagonalBatch.R
\name{multiplyLowerDiagonalBatch}
\alias{multiplyLowerDiagonalBatch}
\title{multiplyLowerDiagonalBatch}
\usage{
multiplyLowerDiagonalBatch(
  L,
  D,
  B,
  output,
  diagIsOne,
  transformD,
  Nglobal,
  Nlocal,
  NlocalCache
)
}
\arguments{
\item{L}{lower triangular matrices in batches}

\item{D}{diagonal matrices in batches, each row contains diagonal elements of D'}

\item{B}{matrices in batches}

\item{output}{the result of LD'B}

\item{diagIsOne}{logical, whether the diagonal of L is one}

\item{transformD}{how to transform D, can be any OpenCL C built-in math function}

\item{Nglobal}{the size of the index space for use}

\item{Nlocal}{the work group size of the index space}

\item{NlocalCache}{a number}
}
\description{
computes output = LD'B in batches on a GPU
}
