% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfLogL.R
\name{ParamsFromLgm}
\alias{ParamsFromLgm}
\title{Betas and sigma automatic configuration}
\usage{
ParamsFromLgm(
  modelSummary,
  covariates,
  Betaleft = 0.5,
  Betaright = 0.3,
  sdSpatialleft = 0.9,
  sdSpatialright = 1.05
)
}
\description{
Betas and sigma automatic configuration
}
