% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prof2dBetas.R
\name{Prof2dBetas}
\alias{Prof2dBetas}
\title{Estimate Log-likelihood for Gaussian random fields when Betas are given}
\usage{
Prof2dBetas(
  Betas,
  prof2list,
  Nobs,
  Nparam,
  Ndata,
  detVar,
  ssqY,
  XVYXVX,
  jacobian,
  cilevel
)
}
\description{
Estimate Log-likelihood for Gaussian random fields when Betas are given
}
