% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfVariance.R
\name{profVariance}
\alias{profVariance}
\title{Estimate Log-likelihood for Gaussian random fields when sdSpatial are given}
\usage{
profVariance(
  sdSpatial,
  cilevel = 0.95,
  Nobs,
  Ndata,
  Nparam,
  Ncov,
  detVar,
  detReml,
  ssqResidual,
  jacobian,
  reml = FALSE
)
}
\description{
Estimate Log-likelihood for Gaussian random fields when sdSpatial are given
}
